make position improvement offset a parameter; move utililty functions
