+++ .gitignore
@@ -2,4 +2,4 @@
 node_modules
 /.idea
 /examples/*/index.js
-
+/tests/env/index.js
